move run_deployment and load_flow_from_flow_run tests, remove other deployment tests
